+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## CHANGELOG
 
+### UNRELEASED CHANGES
+- Fix: Fix our `APP_CONFIG` mocking in storybook so that `showMedium` and `companionApp` can be set via the toolbar once again.
+
 ### 4.0.0
 - BREAKING CHANGE: Start app by fetching auth document first instead of catalog root. This means that the `libraries` key in the config file needs to be set with auth document urls instead of catalog root urls.
 - Feature: Support auth-protected collections.